@@ -22,54 +22,54 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/config": "^2.0.1",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/jwt": "^8.0.1",
-    "@nestjs/microservices": "^8.4.5",
-    "@nestjs/passport": "^8.2.1",
-    "@nestjs/platform-express": "^8.0.0",
-    "amqp-connection-manager": "^4.1.3",
-    "amqplib": "^0.9.0",
+    "@nestjs/common": "^9.0.9",
+    "@nestjs/config": "^2.2.0",
+    "@nestjs/core": "^9.0.9",
+    "@nestjs/jwt": "^9.0.0",
+    "@nestjs/microservices": "^9.0.9",
+    "@nestjs/passport": "^9.0.0",
+    "@nestjs/platform-express": "^9.0.9",
+    "amqp-connection-manager": "^4.1.6",
+    "amqplib": "^0.10.2",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-plugin-airbnb": "^0.0.1-security",
-    "eslint-plugin-unicorn": "^42.0.0",
+    "eslint-plugin-unicorn": "^43.0.2",
     "joi": "^17.6.0",
-    "passport": "^0.5.3",
+    "passport": "^0.6.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "^7.2.0"
+    "rxjs": "^7.5.6"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.0.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
+    "@nestjs/cli": "^9.0.0",
+    "@nestjs/schematics": "^9.0.1",
+    "@nestjs/testing": "^9.0.9",
     "@types/express": "^4.17.13",
-    "@types/jest": "27.4.1",
-    "@types/node": "^16.0.0",
+    "@types/jest": "28.1.6",
+    "@types/node": "^18.7.2",
     "@types/passport-jwt": "^3.0.6",
     "@types/passport-local": "^1.0.34",
-    "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^8.0.1",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^5.33.0",
+    "@typescript-eslint/parser": "^5.33.0",
+    "eslint": "^8.21.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.2.5",
-    "prettier": "^2.3.2",
-    "source-map-support": "^0.5.20",
-    "supertest": "^6.1.3",
-    "ts-jest": "^27.0.3",
-    "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "^3.10.1",
-    "typescript": "^4.3.5",
-    "husky": "^8.0.0"
+    "eslint-plugin-prettier": "^4.2.1",
+    "husky": "^8.0.1",
+    "jest": "^28.1.3",
+    "prettier": "^2.7.1",
+    "source-map-support": "^0.5.21",
+    "supertest": "^6.2.4",
+    "ts-jest": "^28.0.7",
+    "ts-loader": "^9.3.1",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.1.0",
+    "typescript": "^4.7.4"
   },
   "jest": {
     "moduleFileExtensions": [
