@@ -4,4 +4,10 @@ import { SendResetCodeDto } from '../../password/dto/SendResetCodeDto';
 import { GetTokenDto } from '../../password/dto/GetTokenDto';
 import { ResetPasswordDto } from '../../password/dto/ResetPasswordDto';
 
-export type AuthServiceDto = ConfirmCodeDto | UserCreateDto | SendResetCodeDto | GetTokenDto | ResetPasswordDto;
+export type AuthServiceDto =
+    | ConfirmCodeDto
+    | UserCreateDto
+    | SendResetCodeDto
+    | GetTokenDto
+    | ResetPasswordDto
+    | Record<string, never>;
