@@ -6,22 +6,26 @@ import { AppModule } from './app.module';
 async function bootstrap(): Promise<void> {
     const app: INestApplication = await NestFactory.create(AppModule);
 
+    app.useGlobalPipes(
+        new ValidationPipe({
+            whitelist: true,
+            transform: true,
+        }),
+    );
+    app.enableCors();
     app.connectMicroservice({
         transport: Transport.RMQ,
         options: {
-            urls: ['amqp://rabbitmq:5672'], // amqp://rabbitmq:5672// amqp://localhost:5672
+            urls: ['amqp://localhost:5672'], // amqp://rabbitmq:5672// amqp://localhost:5672
             queueOptions: {
                 durable: false,
             },
         },
     });
-    app.useGlobalPipes(
-        new ValidationPipe({
-            whitelist: true,
-        }),
-    );
+
     await app.startAllMicroservices();
-    await app.listen(3000); // new ConfigService().get('port')
+    await app.listen(3001);
 }
 
+/* eslint-disable unicorn/prefer-top-level-await */
 bootstrap();
