@@ -1,4 +1,4 @@
-import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
+import { IsEmail, IsNotEmpty, IsString, MaxLength, MinLength } from 'class-validator';
 
 export class ResetPasswordDto {
     @IsEmail()
@@ -10,6 +10,8 @@ export class ResetPasswordDto {
     token: string;
 
     @IsString()
+    @MinLength(8)
+    @MaxLength(20)
     @IsNotEmpty()
     password: string;
 }
