@@ -1,13 +1,19 @@
 import { Response } from 'express';
 import { HttpStatus } from '@nestjs/common';
 
-interface IErrorHandler {
-    handleErrors: (r: Response, e: Error) => Response;
+interface ServiceError extends Error {
+    code?: number;
 }
-export const ErrorHandler: IErrorHandler = {
-    handleErrors(res: Response, e: Error): Response {
-        const { message } = e;
 
-        return res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({ message });
+interface IErrorHandler {
+    handleErrors: (r: Response, e: ServiceError) => Response;
+}
+
+export const ErrorHandler: IErrorHandler = {
+    handleErrors(res: Response, e: ServiceError): Response {
+        const { message, code } = e;
+        const statusCode: number = code ?? HttpStatus.INTERNAL_SERVER_ERROR;
+
+        return res.status(statusCode).send({ message });
     },
 };
